去手續費

+++ network.js
@@ -50,6 +50,7 @@ var peer_events_buffer = [];
 var assocKnownPeers = {};
 var assocBlockedPeers = {};
 var exchangeRates = {};
+var bWatchingForLight = false;
 
 if (process.browser){ // browser
 	console.log("defining .on() on ws");
@@ -617,7 +618,7 @@ function handleNewPeers(ws, request, arrPeerUrls){
 		var url = arrPeerUrls[i];
 		if (conf.myUrl && conf.myUrl.toLowerCase() === url.toLowerCase())
 			continue;
-		var regexp = (conf.WS_PROTOCOL === 'wss://') ? /^wss:\/\// : /^wss?:\/\//;
+		var regexp = (conf.WS_PROTOCOL === 'wss://') ? /^wss:\/\// : /^ws?:\/\//;
 		if (!url.match(regexp)){
 			console.log('ignoring new peer '+url+' because of incompatible ws protocol');
 			continue;
@@ -873,9 +874,7 @@ function havePendingJointRequest(unit){
 function purgeJunkUnhandledJoints(){
 	if (bCatchingUp || Date.now() - coming_online_time < 3600*1000 || wss.clients.length === 0 && arrOutboundPeers.length === 0)
 		return;
-	db.query("DELETE FROM unhandled_joints WHERE creation_date < "+db.addTime("-1 HOUR"), function(){
-		db.query("DELETE FROM dependencies WHERE NOT EXISTS (SELECT * FROM unhandled_joints WHERE unhandled_joints.unit=dependencies.unit)");
-	});
+	joint_storage.purgeOldUnhandledJoints();
 }
 
 function purgeJointAndDependenciesAndNotifyPeers(objJoint, error, onDone){
@@ -944,13 +943,9 @@ function handleJoint(ws, objJoint, bSaved, callbacks){
 					callbacks.ifJointError(error);
 				//	throw Error(error);
 					unlock();
-					db.query(
-						"INSERT INTO known_bad_joints (joint, json, error) VALUES (?,?,?)", 
-						[objectHash.getJointHash(objJoint), JSON.stringify(objJoint), error],
-						function(){
-							delete assocUnitsInWork[unit];
-						}
-					);
+					joint_storage.saveKnownBadJoint(objJoint, error, function(){
+						delete assocUnitsInWork[unit];
+					});
 					if (ws)
 						writeEvent('invalid', ws.host);
 					if (objJoint.unsigned)
@@ -1269,6 +1264,8 @@ function notifyWatchers(objJoint, source_ws){
 		return;
 	if (objJoint.ball) // already stable, light clients will require a proof
 		return;
+	if (!bWatchingForLight)
+		return;
 	// this is a new unstable joint, light clients will accept it without proof
 	db.query("SELECT peer FROM watched_light_addresses WHERE address IN(?)", [arrAddresses], function(rows){
 		if (rows.length === 0)
@@ -1292,7 +1289,7 @@ function notifyWatchersAboutStableJoints(mci){
 		unlock(); // we don't need to block writes, we requested the lock just to wait that the current write completes
 		notifyLocalWatchedAddressesAboutStableJoints(mci);
 		console.log("notifyWatchersAboutStableJoints "+mci);
-		if (mci <= 1)
+		if (mci <= 1 || !bWatchingForLight)
 			return;
 		storage.findLastBallMciOfMci(db, mci, function(last_ball_mci){
 			storage.findLastBallMciOfMci(db, mci-1, function(prev_last_ball_mci){
@@ -2284,6 +2281,7 @@ function handleJustsaying(ws, subject, body){
 			var address = body;
 			if (!ValidationUtils.isValidAddress(address))
 				return sendError(ws, "address not valid");
+			bWatchingForLight = true;
 			db.query("INSERT "+db.getIgnore()+" INTO watched_light_addresses (peer, address) VALUES (?,?)", [ws.peer, address], function(){
 				sendInfo(ws, "now watching "+address);
 				// check if we already have something on this address
@@ -2326,7 +2324,7 @@ function handleJustsaying(ws, subject, body){
 			if (!ws.bLoggingIn && !ws.bLoggedIn) // accept from hub only
 				return;
 			ws.close(1000, "my core is old");
-			throw Error("Mandatory upgrade required, please check the release notes at https://github.com/byteball/byteball/releases and upgrade.");
+			throw Error("Mandatory upgrade required, please check the release notes at https://github.com/luxalpa/luxalpa/releases and upgrade.");
 			break;
 			
 		case 'custom':
@@ -2581,6 +2579,7 @@ function handleRequest(ws, tag, command, params){
 					},
 					ifOk: function(objResponse){
 						sendResponse(ws, tag, objResponse);
+						bWatchingForLight = true;
 						if (params.addresses)
 							db.query(
 								"INSERT "+db.getIgnore()+" INTO watched_light_addresses (peer, address) VALUES "+
@@ -2944,9 +2943,11 @@ function startAcceptingConnections(){
 			tryHandleMessage();
 		});
 		ws.on('close', function(){
-			db.query("DELETE FROM watched_light_addresses WHERE peer=?", [ws.peer]);
-			db.query("DELETE FROM watched_light_units WHERE peer=?", [ws.peer]);
-			//db.query("DELETE FROM light_peer_witnesses WHERE peer=?", [ws.peer]);
+			if (bWatchingForLight){
+				db.query("DELETE FROM watched_light_addresses WHERE peer=?", [ws.peer]);
+				db.query("DELETE FROM watched_light_units WHERE peer=?", [ws.peer]);
+				//db.query("DELETE FROM light_peer_witnesses WHERE peer=?", [ws.peer]);
+			}
 			console.log("client "+ws.peer+" disconnected");
 			cancelRequestsOnClosedConnection(ws);
 		});
@@ -2966,6 +2967,7 @@ function startRelay(){
 		startAcceptingConnections();
 	
 	storage.initCaches();
+	joint_storage.initUnhandledAndKnownBad();
 	checkCatchupLeftovers();
 
 	if (conf.bWantNewPeers){
